changed parameter double to int

+++ mobile-app/app/src/main/java/com/example/ballbouncermobilapp/DataFromServer.java
@@ -3,22 +3,22 @@ package com.example.ballbouncermobilapp;
 import java.io.Serializable;
 
 public class DataFromServer implements Serializable {
-    private int x;
-    private int y;
-    private int z;
-    private int _x;
-    private int _y;
-    private int _z;
+    private double x;
+    private double y;
+    private double z;
+    private double _x;
+    private double _y;
+    private double _z;
 
     public DataFromServer() {
     }
 
-
-    public void generalSetter(int px, int py, int pz){
+    public void generalSetter(double px, double py, double pz){
 
         this.x = px;
         this.y = py;
         this.z = pz;
+
         if(this.z == 0.0) {
             this._x = this.x;
             this._y = this.y;
@@ -26,51 +26,51 @@ public class DataFromServer implements Serializable {
         }
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(double z) {
         this.z = z;
     }
 
-    public int get_x() {
+    public double get_x() {
         return _x;
     }
 
-    public void set_x(int _x) {
+    public void set_x(double _x) {
         this._x = _x;
     }
 
-    public int get_y() {
+    public double get_y() {
         return _y;
     }
 
-    public void set_y(int _y) {
+    public void set_y(double _y) {
         this._y = _y;
     }
 
-    public int get_z() {
+    public double get_z() {
         return _z;
     }
 
-    public void set_z(int _z) {
+    public void set_z(double _z) {
         this._z = _z;
     }
 }
